Updated product details screen(favorites, navigation, ui) - Favorites integration - Navigation to search, sub products details - Fix display article number and comments

+++ app/src/main/java/com/vodovoz/app/data/vodovoz_service/model/POPULAR_CATEGORY_DTO.kt
@@ -7,7 +7,7 @@ import com.squareup.moshi.Json
 @Keep
 data class POPULAR_CATEGORY_DTO(
     @Json(name = "IDRAZDEL")
-    val IDRAZDEL: Int?,
+    val IDRAZDEL: Long?,
     @Json(name = "NAMERAZDEL")
     val NAMERAZDEL: String?,
     @Json(name = "PICTURE")
