+++ app/src/main/java/com/vodovoz/app/data/vodovoz_service/model/catalog_details/CATALOG_CATEGORY_DTO.kt
@@ -11,7 +11,7 @@ data class CATALOG_CATEGORY_DTO(
     @Json(name = "IBLOCK_SECTION_ID")
     val IBLOCK_SECTION_ID: Int?,
     @Json(name = "ID")
-    val ID: Int?,
+    val ID: Long?,
     @Json(name = "NAME")
     val NAME: String?,
     @Json(name = "PICTURE")
